// EcomWebsiteFE/src/types/types.ts
export interface IContactItems {
    id: number;
    title: string;
    content: JSX.Element;
};